move account deletion to new file
