+++ wire-ios-sync-engine/Tests/Source/SessionManager/SessionManagerTests.swift
@@ -457,75 +457,6 @@ extension IntegrationTest {
 
 // MARK: - Account Deletion
 
-class SessionManagertests_AccountDeletion: IntegrationTest {
-
-    override func setUp() {
-        super.setUp()
-        createSelfUserAndConversation()
-    }
-
-    func testThatItDeletesTheAccountFolder_WhenDeletingAccountWithoutActiveUserSession() throws {
-        // given
-        guard let sharedContainer = Bundle.main.appGroupIdentifier.map(FileManager.sharedContainerDirectory) else {
-            XCTFail()
-            return
-        }
-        let account = self.createAccount()
-
-        let accountFolder = CoreDataStack.accountDataFolder(accountIdentifier: account.userIdentifier, applicationContainer: sharedContainer)
-
-        try FileManager.default.createDirectory(at: accountFolder, withIntermediateDirectories: true, attributes: nil)
-
-        // when
-        performIgnoringZMLogError {
-            self.sessionManager!.delete(account: account)
-        }
-
-        // then
-        XCTAssertFalse(FileManager.default.fileExists(atPath: accountFolder.path))
-    }
-
-    func testThatItDeletesTheAccountFolder_WhenDeletingActiveUserSessionAccount() throws {
-        // given
-        XCTAssert(login())
-
-        guard let sharedContainer = Bundle.main.appGroupIdentifier.map(FileManager.sharedContainerDirectory) else { return XCTFail() }
-
-        let account = sessionManager!.accountManager.selectedAccount!
-        let accountFolder = CoreDataStack.accountDataFolder(accountIdentifier: account.userIdentifier, applicationContainer: sharedContainer)
-
-        // when
-        performIgnoringZMLogError {
-            self.sessionManager!.delete(account: account)
-        }
-
-        // then
-        XCTAssertFalse(FileManager.default.fileExists(atPath: accountFolder.path))
-    }
-
-    func testThatItDeletesTheLastEventID_WhenDeletingActiveUserSessionAccount() throws {
-        // given
-        XCTAssert(login())
-
-        let sessionManager = try XCTUnwrap(sessionManager)
-        let account = try XCTUnwrap(sessionManager.accountManager.selectedAccount)
-        let repository = LastEventIDRepository(
-            userID: account.userIdentifier,
-            sharedUserDefaults: sharedUserDefaults
-        )
-        XCTAssertNotNil(repository.fetchLastEventID())
-
-        // when
-        performIgnoringZMLogError {
-            sessionManager.delete(account: account)
-        }
-
-        // then
-        XCTAssertNil(repository.fetchLastEventID())
-    }
-
-}
-
 class SessionManagerTests_AuthenticationFailure: IntegrationTest {
 
     override func setUp() {
